refactor+feat: from `slog+hyper` to `tracing+axum` This is a massive refactor which will enable us to collect better logs. Moves code out of main, so now the bulk of the proxy is in `proxy.rs`. Adds a new log format switch with `json` support. The exact format may need some fine tuning.
