+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "icx-proxy"
-version = "0.9.0"
+version = "0.10.0"
 authors = ["DFINITY Stiftung <sdk@dfinity.org>"]
 edition = "2018"
 description = "CLI tool to create an HTTP proxy to the Internet Computer."
@@ -18,7 +18,7 @@ path = "src/main.rs"
 
 [dependencies]
 anyhow = "1"
-axum = "0.5.3"
+axum = "0.5.10"
 base64 = "0.13"
 candid = { version = "0.7", features = ["mute_warnings"] }
 clap = { version = "3", features = ["cargo", "derive"] }
@@ -35,17 +35,17 @@ ic-utils = { version = "0.20", features = ["raw"] }
 lazy-regex = "2"
 opentelemetry = "0.17.0"
 opentelemetry-prometheus = "0.10.0"
-prometheus = "0.13.0"
-reqwest = { version = "0.11", features = ["rustls-tls-webpki-roots"] }
+prometheus = "0.13.1"
+reqwest = { version = "0.11.11", features = ["rustls-tls-webpki-roots"] }
 rustls = "0.20"
 rustls-pemfile = "1"
+tower-http = { version = "0.3", features = ["trace"] }
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["json"]}
 serde = "1"
 serde_cbor = "0.11"
 serde_json = "1"
 sha2 = "0.10"
-slog = { version = "2", features = ["max_level_trace"] }
-slog-async = "2"
-slog-term = "2"
 tokio = { version = "1", features = ["full"] }
 url = "2"
 webpki-roots = "0.22"
