+++ src/metrics.rs
@@ -1,9 +1,28 @@
+use std::{future::Future, net::SocketAddr};
+
+use axum::{handler::Handler, routing::get, Extension, Router};
+use clap::Args;
+use futures::{future::OptionFuture, FutureExt};
+use hyper::{Body, Request, Response, StatusCode};
 use opentelemetry::{
+    global,
     metrics::{Counter, Meter},
+    sdk::Resource,
     KeyValue,
 };
+use opentelemetry_prometheus::PrometheusExporter;
+use prometheus::{Encoder, TextEncoder};
 
-use crate::validate::Validate;
+use crate::{logging::add_trace_layer, validate::Validate};
+
+/// The options for metrics
+#[derive(Args)]
+pub struct Opts {
+    /// Address to expose Prometheus metrics on
+    /// Examples: 127.0.0.1:9090, [::1]:9090
+    #[clap(long)]
+    metrics_addr: Option<SocketAddr>,
+}
 
 pub struct WithMetrics<T>(pub T, pub MetricParams);
 
@@ -30,11 +49,10 @@ impl<T: Validate> Validate for WithMetrics<T> {
         agent: &ic_agent::Agent,
         uri: &hyper::Uri,
         response_body: &[u8],
-        logger: slog::Logger,
     ) -> Result<(), String> {
         let out = self
             .0
-            .validate(headers_data, canister_id, agent, uri, response_body, logger);
+            .validate(headers_data, canister_id, agent, uri, response_body);
 
         let mut status = if out.is_ok() { "ok" } else { "fail" };
         if cfg!(feature = "skip_body_verification") {
@@ -47,5 +65,63 @@ impl<T: Validate> Validate for WithMetrics<T> {
         counter.add(1, labels);
 
         out
+    }
+}
+
+#[derive(Clone)]
+struct HandlerArgs {
+    exporter: PrometheusExporter,
+}
+
+async fn metrics_handler(
+    Extension(HandlerArgs { exporter }): Extension<HandlerArgs>,
+    _: Request<Body>,
+) -> Response<Body> {
+    let metric_families = exporter.registry().gather();
+
+    let encoder = TextEncoder::new();
+
+    let mut metrics_text = Vec::new();
+    if encoder.encode(&metric_families, &mut metrics_text).is_err() {
+        return Response::builder()
+            .status(StatusCode::INTERNAL_SERVER_ERROR)
+            .body("Internal Server Error".into())
+            .unwrap();
+    };
+
+    Response::builder()
+        .status(200)
+        .body(metrics_text.into())
+        .unwrap()
+}
+pub fn setup(opts: Opts) -> (Meter, Runner) {
+    let exporter = opentelemetry_prometheus::exporter()
+        .with_resource(Resource::new(vec![KeyValue::new("service", "prober")]))
+        .init();
+    (
+        global::meter("icx-proxy"),
+        Runner {
+            exporter,
+            metrics_addr: opts.metrics_addr,
+        },
+    )
+}
+
+pub struct Runner {
+    exporter: PrometheusExporter,
+    metrics_addr: Option<SocketAddr>,
+}
+
+impl Runner {
+    pub fn run(self) -> impl Future<Output = Result<Option<()>, hyper::Error>> {
+        let exporter = self.exporter;
+        OptionFuture::from(self.metrics_addr.map(|metrics_addr| {
+            let metrics_handler = metrics_handler.layer(Extension(HandlerArgs { exporter }));
+            let metrics_router = Router::new().route("/metrics", get(metrics_handler));
+
+            axum::Server::bind(&metrics_addr)
+                .serve(add_trace_layer(metrics_router).into_make_service())
+        }))
+        .map(|v| v.transpose())
     }
 }
